agent/pyproject.toml
[build-system]
requires = [
  "setuptools>=61.0",
  "wheel"
]
build-backend = "setuptools.build_meta"

[project]
name = "agent-starter-python"
version = "1.0.0"
description = "Simple voice AI assistant built with LiveKit Agents for Python"
requires-python = ">=3.10, <3.15"
dependencies = [
  "httpx[socks]>=0.28.1",
  "livekit-agents[silero,turn-detector,openai,elevenlabs]~=1.5",
  "livekit-plugins-ai-coustics~=0.2",
  "livekit-plugins-minimax-ai>=1.2.17",
  "livekit-plugins-noise-cancellation~=0.2",
  "python-dotenv",
]

[dependency-groups]
dev = [
  "pytest",
  "pytest-asyncio",
  "ruff",
]

[tool.setuptools.packages.find]
where = [ "src" ]

[tool.setuptools.package-dir]
"" = "src"

[tool.pytest.ini_options]
asyncio_mode = "auto"
asyncio_default_fixture_loop_scope = "function"

[tool.ruff]
line-length = 88
target-version = "py39"

[tool.ruff.lint]
select = [
  "E",
  "F",
  "W",
  "I",
  "N",
  "B",
  "A",
  "C4",
  "UP",
  "SIM",
  "RUF"
]
ignore = [ "E501", "RUF001" ] # Line too long (handled by formatter); allow fullwidth chars in Chinese prompts

[tool.ruff.format]
quote-style = "double"
indent-style = "space"
